implement tunnel-stats for macos
