+++ src-tauri/src/apple.rs
@@ -1,28 +1,13 @@
 //! Interchangeability and communication with VPNExtension (written in Swift).
 
-use std::{
-    collections::HashMap,
-    hint::spin_loop,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc, Mutex,
-    },
-    time::Duration,
-};
+use std::{collections::HashMap, time::Duration};
 
-use block2::RcBlock;
 use defguard_client_core::connection::{
     active_connections::find_connection,
-    apple::{
-        manager_for_key_and_value, LOCATION_ID, PLUGIN_BUNDLE_ID, TUNNEL_ID, VPN_STATE_UPDATE_COMMS,
-    },
+    apple::{manager_for_key_and_value, LOCATION_ID, TUNNEL_ID, VPN_STATE_UPDATE_COMMS},
 };
 use objc2::rc::Retained;
-use objc2_foundation::{ns_string, NSData};
-use objc2_network_extension::{
-    NETunnelProviderManager, NETunnelProviderProtocol, NETunnelProviderSession, NEVPNStatus,
-};
-use serde::Deserialize;
+use objc2_network_extension::{NETunnelProviderManager, NEVPNStatus};
 use tauri::{AppHandle, Emitter, Manager};
 use tokio::time::sleep;
 use tracing::Level;
@@ -265,18 +250,6 @@ async fn sync_connections_with_system(app_handle: &AppHandle) {
     }
 }
 
-/// Tunnel statistics shared with VPNExtension (written in Swift).
-#[derive(Deserialize)]
-#[repr(C)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct Stats {
-    pub(crate) location_id: Option<Id>,
-    pub(crate) tunnel_id: Option<Id>,
-    pub(crate) tx_bytes: u64,
-    pub(crate) rx_bytes: u64,
-    pub(crate) last_handshake: u64,
-}
-
 #[must_use]
 pub fn get_managers_for_tunnels_and_locations(
     tunnels: &[Tunnel<Id>],
@@ -297,80 +270,4 @@ pub fn get_managers_for_tunnels_and_locations(
     }
 
     managers
-}
-
-/// Retrieve VPN tunnel statistics from VPNExtension.
-pub(crate) fn tunnel_stats(id: Id, connection_type: &ConnectionType) -> Option<Stats> {
-    let new_stats = Arc::new(Mutex::new(None));
-    let plugin_bundle_id = ns_string!(PLUGIN_BUNDLE_ID);
-
-    let new_stats_clone = Arc::clone(&new_stats);
-
-    let finished = Arc::new(AtomicBool::new(false));
-    let finished_clone = Arc::clone(&finished);
-
-    let response_handler = RcBlock::new(move |data_ptr: *mut NSData| {
-        if let Some(data) = unsafe { data_ptr.as_ref() } {
-            if let Ok(stats) = serde_json::from_slice(data.to_vec().as_slice()) {
-                if let Ok(mut new_stats_locked) = new_stats_clone.lock() {
-                    *new_stats_locked = Some(stats);
-                }
-            } else {
-                warn!("Failed to deserialize tunnel stats");
-            }
-        } else {
-            debug!("No data received in tunnel stats response, skipping");
-        }
-        finished_clone.store(true, Ordering::Release);
-    });
-
-    let manager = manager_for_key_and_value(
-        match connection_type {
-            ConnectionType::Location => LOCATION_ID,
-            ConnectionType::Tunnel => TUNNEL_ID,
-        },
-        id,
-    )?;
-
-    let vpn_protocol = (unsafe { manager.protocolConfiguration() })?;
-    let Ok(tunnel_protocol) = vpn_protocol.downcast::<NETunnelProviderProtocol>() else {
-        error!("Failed to downcast to NETunnelProviderProtocol");
-        return None;
-    };
-
-    // Sometimes all managers from all apps come through, so filter by bundle ID.
-    if let Some(bundle_id) = unsafe { tunnel_protocol.providerBundleIdentifier() } {
-        if &*bundle_id != plugin_bundle_id {
-            return None;
-        }
-    }
-
-    let Ok(session) = unsafe { manager.connection() }.downcast::<NETunnelProviderSession>() else {
-        error!("Failed to downcast to NETunnelProviderSession");
-        return None;
-    };
-
-    let message_data = NSData::new();
-    if unsafe {
-        session.sendProviderMessage_returnError_responseHandler(
-            &message_data,
-            None,
-            Some(&response_handler),
-        )
-    } {
-        debug!("Message sent to NETunnelProviderSession");
-    } else {
-        error!("Failed to send to NETunnelProviderSession while requesting stats");
-    }
-
-    // Wait for all handlers to complete.
-    while !finished.load(Ordering::Acquire) {
-        spin_loop();
-    }
-
-    let stats = new_stats
-        .lock()
-        .map_or(None, |mut new_stats_locked| new_stats_locked.take());
-
-    stats
 }
